Remove incorrect user-script FAQ entry for publishing JSON data
This was a carry-over from a webviz feature that has not existed in Foxglove for a long time (years?). This change removes this incorrect documentation. This removal left one item in the FAQ section so I remove "FAQ" and made the remaining item take its place in the subheaders heirarchy.