Document some limits for free plans (#179)
This makes note of a couple of limitations in the new free plan, which
is now in effect (https://github.com/foxglove/app/pull/5575).

The new free plan is described by the pricing updates in
https://github.com/foxglove/website/pull/1244 with the latest preview at
https://website-g73j8cil3.foxglove.party/pricing.

Free plans are also now granted storage and can upload & stream. I don't
think this restriction was noted in the documentation before, because we
start orgs on the "team trial" plan today.